+++ core/tests/network_dial_error.rs
@@ -22,47 +22,60 @@ mod util;
 
 use futures::prelude::*;
 use libp2p_core::identity;
-use libp2p_core::multiaddr::multiaddr;
+use libp2p_core::multiaddr::{multiaddr, Multiaddr};
 use libp2p_core::{
     Network,
     PeerId,
     Transport,
     connection::PendingConnectionError,
     muxing::StreamMuxerBox,
-    network::NetworkEvent,
+    network::{NetworkEvent, NetworkConfig},
+    transport,
     upgrade,
 };
+use rand::Rng;
 use rand::seq::SliceRandom;
-use std::{io, task::Poll};
+use std::{io, error::Error, fmt, task::Poll};
 use util::TestHandler;
 
-type TestNetwork<TTrans> = Network<TTrans, (), (), TestHandler>;
+type TestNetwork = Network<TestTransport, (), (), TestHandler>;
+type TestTransport = transport::boxed::Boxed<(PeerId, StreamMuxerBox), BoxError>;
+
+#[derive(Debug)]
+struct BoxError(Box<dyn Error + Send + 'static>);
+
+impl Error for BoxError {}
+
+impl fmt::Display for BoxError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "Transport error: {}", self.0)
+    }
+}
+
+fn new_network(cfg: NetworkConfig) -> TestNetwork {
+    let local_key = identity::Keypair::generate_ed25519();
+    let local_public_key = local_key.public();
+    let transport: TestTransport = libp2p_tcp::TcpConfig::new()
+        .upgrade(upgrade::Version::V1)
+        .authenticate(libp2p_secio::SecioConfig::new(local_key))
+        .multiplex(libp2p_mplex::MplexConfig::new())
+        .map(|(conn_info, muxer), _| (conn_info, StreamMuxerBox::new(muxer)))
+        .and_then(|(peer, mplex), _| {
+            // Gracefully close the connection to allow protocol
+            // negotiation to complete.
+            util::CloseMuxer::new(mplex).map_ok(move |mplex| (peer, mplex))
+        })
+        .map_err(|e| BoxError(Box::new(e)))
+        .boxed();
+    TestNetwork::new(transport, local_public_key.into(), cfg)
+}
 
 #[test]
 fn deny_incoming_connec() {
     // Checks whether refusing an incoming connection on a swarm triggers the correct events.
 
-    let mut swarm1 = {
-        let local_key = identity::Keypair::generate_ed25519();
-        let local_public_key = local_key.public();
-        let transport = libp2p_tcp::TcpConfig::new()
-            .upgrade(upgrade::Version::V1)
-            .authenticate(libp2p_secio::SecioConfig::new(local_key))
-            .multiplex(libp2p_mplex::MplexConfig::new())
-            .map(|(conn_info, muxer), _| (conn_info, StreamMuxerBox::new(muxer)));
-        TestNetwork::new(transport, local_public_key.into(), Default::default())
-    };
-
-    let mut swarm2 = {
-        let local_key = identity::Keypair::generate_ed25519();
-        let local_public_key = local_key.public();
-        let transport = libp2p_tcp::TcpConfig::new()
-            .upgrade(upgrade::Version::V1)
-            .authenticate(libp2p_secio::SecioConfig::new(local_key))
-            .multiplex(libp2p_mplex::MplexConfig::new())
-            .map(|(conn_info, muxer), _| (conn_info, StreamMuxerBox::new(muxer)));
-        TestNetwork::new(transport, local_public_key.into(), Default::default())
-    };
+    let mut swarm1 = new_network(NetworkConfig::default());
+    let mut swarm2 = new_network(NetworkConfig::default());
 
     swarm1.listen_on("/ip4/127.0.0.1/tcp/0".parse().unwrap()).unwrap();
 
@@ -76,8 +89,7 @@ fn deny_incoming_connec() {
 
     swarm2
         .peer(swarm1.local_peer_id().clone())
-        .into_disconnected().unwrap()
-        .connect(address.clone(), Vec::new(), TestHandler())
+        .dial(address.clone(), Vec::new(), TestHandler())
         .unwrap();
 
     async_std::task::block_on(future::poll_fn(|cx| -> Poll<Result<(), io::Error>> {
@@ -119,22 +131,7 @@ fn dial_self() {
     //
     // The last two can happen in any order.
 
-    let mut swarm = {
-        let local_key = identity::Keypair::generate_ed25519();
-        let local_public_key = local_key.public();
-        let transport = libp2p_tcp::TcpConfig::new()
-            .upgrade(upgrade::Version::V1)
-            .authenticate(libp2p_secio::SecioConfig::new(local_key))
-            .multiplex(libp2p_mplex::MplexConfig::new())
-            .and_then(|(peer, mplex), _| {
-                // Gracefully close the connection to allow protocol
-                // negotiation to complete.
-                util::CloseMuxer::new(mplex).map_ok(move |mplex| (peer, mplex))
-            })
-            .map(|(conn_info, muxer), _| (conn_info, StreamMuxerBox::new(muxer)));
-        TestNetwork::new(transport, local_public_key.into(), Default::default())
-    };
-
+    let mut swarm = new_network(NetworkConfig::default());
     swarm.listen_on("/ip4/127.0.0.1/tcp/0".parse().unwrap()).unwrap();
 
     let (local_address, mut swarm) = async_std::task::block_on(
@@ -193,36 +190,16 @@ fn dial_self() {
 fn dial_self_by_id() {
     // Trying to dial self by passing the same `PeerId` shouldn't even be possible in the first
     // place.
-
-    let mut swarm = {
-        let local_key = identity::Keypair::generate_ed25519();
-        let local_public_key = local_key.public();
-        let transport = libp2p_tcp::TcpConfig::new()
-            .upgrade(upgrade::Version::V1)
-            .authenticate(libp2p_secio::SecioConfig::new(local_key))
-            .multiplex(libp2p_mplex::MplexConfig::new())
-            .map(|(conn_info, muxer), _| (conn_info, StreamMuxerBox::new(muxer)));
-        TestNetwork::new(transport, local_public_key.into(), Default::default())
-    };
-
+    let mut swarm = new_network(NetworkConfig::default());
     let peer_id = swarm.local_peer_id().clone();
     assert!(swarm.peer(peer_id).into_disconnected().is_none());
 }
 
 #[test]
 fn multiple_addresses_err() {
-    // Tries dialing multiple addresses, and makes sure there's one dialing error per addresses.
+    // Tries dialing multiple addresses, and makes sure there's one dialing error per address.
 
-    let mut swarm = {
-        let local_key = identity::Keypair::generate_ed25519();
-        let local_public_key = local_key.public();
-        let transport = libp2p_tcp::TcpConfig::new()
-            .upgrade(upgrade::Version::V1)
-            .authenticate(libp2p_secio::SecioConfig::new(local_key))
-            .multiplex(libp2p_mplex::MplexConfig::new())
-            .map(|(conn_info, muxer), _| (conn_info, StreamMuxerBox::new(muxer)));
-        TestNetwork::new(transport, local_public_key.into(), Default::default())
-    };
+    let mut swarm = new_network(NetworkConfig::default());
 
     let mut addresses = Vec::new();
     for _ in 0 .. 3 {
@@ -238,8 +215,7 @@ fn multiple_addresses_err() {
 
     let target = PeerId::random();
     swarm.peer(target.clone())
-        .into_disconnected().unwrap()
-        .connect(first, rest, TestHandler())
+        .dial(first, rest, TestHandler())
         .unwrap();
 
     async_std::task::block_on(future::poll_fn(|cx| -> Poll<Result<(), io::Error>> {
@@ -266,4 +242,45 @@ fn multiple_addresses_err() {
             }
         }
     })).unwrap();
+}
+
+#[test]
+fn connection_limit() {
+    let outgoing_per_peer_limit = rand::thread_rng().gen_range(1, 10);
+    let outgoing_limit = 2 * outgoing_per_peer_limit;
+
+    let mut cfg = NetworkConfig::default();
+    cfg.set_outgoing_per_peer_limit(outgoing_per_peer_limit);
+    cfg.set_outgoing_limit(outgoing_limit);
+    let mut network = new_network(cfg);
+
+    let target = PeerId::random();
+    for _ in 0 .. outgoing_per_peer_limit {
+        network.peer(target.clone())
+            .dial(Multiaddr::empty(), Vec::new(), TestHandler())
+            .ok()
+            .expect("Unexpected connection limit.");
+    }
+
+    let err = network.peer(target)
+        .dial(Multiaddr::empty(), Vec::new(), TestHandler())
+        .expect_err("Unexpected dialing success.");
+
+    assert_eq!(err.current, outgoing_per_peer_limit);
+    assert_eq!(err.limit, outgoing_per_peer_limit);
+
+    let target2 = PeerId::random();
+    for _ in outgoing_per_peer_limit .. outgoing_limit {
+        network.peer(target2.clone())
+            .dial(Multiaddr::empty(), Vec::new(), TestHandler())
+            .ok()
+            .expect("Unexpected connection limit.");
+    }
+
+    let err = network.peer(target2)
+        .dial(Multiaddr::empty(), Vec::new(), TestHandler())
+        .expect_err("Unexpected dialing success.");
+
+    assert_eq!(err.current, outgoing_limit);
+    assert_eq!(err.limit, outgoing_limit);
 }